[pua] force pres_id a unsigned_int during calculation
To avoid any issues due to "singed" overflowing

(cherry picked from commit ac99224ef22ece40e3dcd11417712db84d910bb7)

+++ modules/pua/hash.h
@@ -63,7 +63,7 @@
 #define MAX_FORWARD  70
 
 #define PRES_HASH_ID(presentity) \
-	(HASH_SIZE*(long)presentity->local_index + presentity->hash_index)
+	(unsigned long)(HASH_SIZE*(unsigned long)presentity->local_index + presentity->hash_index)
 
 typedef struct publ
 {
